+++ service/accept_statuss_service.go
@@ -5,7 +5,7 @@ import "github.com/IbnuFarhanS/pinjol/model"
 type AcceptStatusService interface {
 	Save(newAcceptStatus model.AcceptStatus) (model.AcceptStatus, error)
 	Update(updateAcceptStatus model.AcceptStatus) (model.AcceptStatus, error)
-	Delete(id int64) (model.AcceptStatus, error)
-	FindById(id int64) (model.AcceptStatus, error)
+	Delete(id uint) (model.AcceptStatus, error)
+	FindById(id uint) (model.AcceptStatus, error)
 	FindAll() ([]model.AcceptStatus, error)
 }
